+++ osu.Game.Rulesets.Osu/Difficulty/Skills/Speed.cs
@@ -13,7 +13,7 @@ namespace osu.Game.Rulesets.Osu.Difficulty.Skills
     /// <summary>
     /// Represents the skill required to press keys with regards to keeping up with the speed at which objects need to be hit.
     /// </summary>
-    public class Speed : Skill
+    public class Speed : StrainSkill
     {
         private const double single_spacing_threshold = 125;
